Add libft header

+++ IC/00_Libft/part_01/ft_memchr.c
@@ -1,4 +1,6 @@
-void    *ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void    *ft_memchr(const void *s, int c, size_t n)
 {
     
 }
