+++ IC/00_Libft/part_01/ft_memcmp.c
@@ -1,4 +1,6 @@
-int ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int ft_memcmp(const void *s1, const void *s2, size_t n)
 {
     
 }
