+++ IC/00_Libft/part_01/ft_memcpy.c
@@ -1,4 +1,6 @@
-void    *ft_memcpy(void *restrict dst, const void *restrict src, unsigned int n)
+#include "libft.h"
+
+void    *ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
 
 }
