+++ IC/00_Libft/part_01/ft_memmove.c
@@ -1,4 +1,6 @@
-void    *ft_memmove(void *dst, const void *src, unsigned int len)
+#include "libft.h"
+
+void    *ft_memmove(void *dst, const void *src, size_t len)
 {
     
 }
