+++ IC/00_Libft/part_01/ft_strncmp.c
@@ -1,4 +1,6 @@
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned int	idx;
