+++ IC/00_Libft/part_01/ft_strnstr.c
@@ -1,4 +1,6 @@
-char    *ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+#include "libft.h"
+
+char    *ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
     int     		idx;
 	unsigned int	needle_len;
